fix: remove 1:1 service ID to entity mapping
Remove the unique constraint on just the service ID column for entity
mappings, instead making it unique over only entityRef. This changes
how service overrides have to work since service IDs are no longer
an equivalent identifier for an entity, so entity providers can no
longer rely on looking up a service ID to get an override. As such, the
service ID lookup endpoint will now return a 404 instead of attempting
a lookup.

The refresh for old entity mappings is removed as all mappings must
have an associated entity reference and can't change entities. It may
be worth removing the uniqueness constraint around service ID as well,
but we aren't using the database for overrides at this point.

This is related to but likely not a proper fix for #92.