fix: set fallback config if only one account is defined
Under the non-legacy config (i.e., a list of named accounts), the
backend plugin fails to correctly configure a fallback for looking up
services under a default account. This appears to be a result of a bug
in how default accounts are handled in the single-account case, causing
it to set EndpointConfigs.default instead of the fallback config. The
former, EC.default, is used only if a legacy config is used, and the
fallback is used under the multi-account setup.

To fix and uncomplicate this, if there is only one account, set the
only account present's isDefault field to true so the fallback config
access doesn't result in an exception. Always walk the set of accounts,
even if there is only one, and move fallback / default assignment into
the account handling function.

By doing this, the same registration logic is applied regardless of
whether there is one account or multiple accounts. The only distinction
is the the first account being made default if only one is present.